requirements.txt
# Telegram Bot Framework
aiogram==3.8.0

# Pydantic v2 (compatible with Aiogram 3.8.0)
pydantic==2.7.1

# Database (PostgreSQL ORM)
sqlalchemy==2.0.31
asyncpg==0.29.0

# API Clients
openai  # <-- Убираем версию, чтобы установить последнюю
pinecone-client==4.1.2

# Environment Variables & Config
python-dotenv==1.0.1

# Tokenizer for context control
tiktoken==0.7.0

# Error monitoring
sentry-sdk==2.8.0

# Task Scheduler
apscheduler==3.10.4

# For data encryption in DB
sqlalchemy-utils==0.41.1
cryptography==42.0.8

# For retrying failed operations
tenacity==9.1.2